refactor: only keep query logic in hook
